refactor: remove domain and integrate to application
