+++ apps/electron/src/main.ts
@@ -4,8 +4,8 @@ import { MCPServerManager } from "@/main/application/mcp-core/mcp-manager";
 import { MCPHttpServer } from "@/main/application/mcp-core/mcp-manager/http/mcp-http-server";
 import started from "electron-squirrel-startup";
 import { updateElectronApp } from "update-electron-app";
-import { setApplicationMenu } from "@/main/application/ui/menu";
-import { createTray, updateTrayContextMenu } from "@/main/application/ui/tray";
+import { setApplicationMenu } from "@/main/ui/menu";
+import { createTray, updateTrayContextMenu } from "@/main/ui/tray";
 import { importExistingServerConfigurations } from "@/main/application/mcp-core/apps/mcp-config-importer";
 import { getPlatformAPIManager } from "@/main/application/workspace/platform-api-manager";
 import { getWorkspaceService } from "./main/application/workspace/workspace-service";
