+++ apps/electron/src/main/application/agent/deployed/deployed-agent-service.ts
@@ -5,7 +5,7 @@ import {
 import { getDeployedAgentRepository } from "../../../infrastructure/database";
 import { DeployedAgent } from "./deployed-agent";
 import { logError, logInfo } from "@/main/utils/logger";
-import { SingletonService } from "@/main/application/core/singleton-service";
+import { SingletonService } from "@/main/application/singleton-service";
 
 /**
  * デプロイ済みエージェント管理サービス（軽量版）
