+++ apps/electron/src/main/application/base-service.ts
