+++ apps/electron/src/main/application/mcp-core/server/server-service.ts
@@ -1,4 +1,4 @@
-import { SingletonService } from "@/main/application/core/singleton-service";
+import { SingletonService } from "@/main/application/singleton-service";
 import { MCPServer, MCPServerConfig } from "@mcp_router/shared";
 import { logInfo } from "@/main/utils/logger";
 import { getServerRepository } from "../../../infrastructure/database";
