+++ apps/electron/src/main/application/singleton-service.ts
