+++ apps/electron/src/main/ui/menu.ts
